+++ packages/validation/schemas/login/email.ts
@@ -1,17 +1,8 @@
-import { LOGIN_CODE_LENGTH } from "@plutomi/shared";
 import { z } from "zod";
 import { emailSchema } from "../shared";
 
 const baseSchema = z.object({
-  email: emailSchema,
-  code: z
-    .string({
-      required_error: "Login code is required"
-    })
-    .length(
-      LOGIN_CODE_LENGTH,
-      `Login code must be ${LOGIN_CODE_LENGTH} characters`
-    )
+  email: emailSchema
 });
 
 export const UISchema = baseSchema;
