+++ packages/validation/schemas/login/index.ts
@@ -1,1 +1,2 @@
-export * from "./login";
+export * as code from "./code";
+export * as email from "./email";
